move the file about database to the proper folder, database.
