+++ api/attractions_api.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, request, jsonify
 import math
-from database.handle_attraction_data import Handle_DB as handle
+from handle_attraction_data import Handle_DB as handle
 
 error={
         "error":True,
